requirements.txt
tensorflow==2.6.0
keras==2.6.0
Pillow==10.0.0
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
numpy==1.21.2